requirements.txt
altair==4.2.0
annotated-types==0.5.0
anyio==3.7.1
attrs==22.2.0
Augmentor==0.2.10
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
deta==1.1.0
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.100.0
filelock==3.12.3
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.42.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.5
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
networkx==3.1
numpy==1.23.5
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==10.0.1
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyodbc==4.0.34
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
requests==2.28.2
rich==13.2.0
scipy==1.11.2
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
streamlit==1.17.0
sympy==1.12
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.64.1
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
ultralytics==8.0.173
urllib3==1.26.14
uvicorn==0.23.1
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.4
win10toast==0.9
WTForms==3.0.1
zipp==3.11.0
